+++ docs/examples/classification/plot_shapelet_locations.py
@@ -3,37 +3,35 @@
 Aligning discovered shapelets with timeseries
 =============================================
 
-This example illustrates the use of the "Learning Shapelets" method in order
-to learn a collection of shapelets that linearly separates the timeseries.
+This example illustrates the use of the "Learning Shapelets" method, presented
+in [1]_ , in order to learn a collection of shapelets that linearly separates
+the timeseries.
 In this example, we will extract a single shapelet in order to distinguish
 between two classes of the "Trace" dataset. Afterwards, we show how our time
 series can be transformed to distances by aligning the shapelets along each of
 the time series. This alignment is performed by shifting the smaller shapelet
 across the longer time series and taking the minimal pointwise distance.
 
-More information on the method can be found at:
-http://fs.ismll.de/publicspace/LearningShapelets/.
+References
+----------
+.. [1] J. Grabocka et al. Learning Time-Series Shapelets. SIGKDD 2014.
 """
-
-# %%
-# .. warning::
-#     When using tensorflow 2.11 to 2.15 (highly probable with python3.8) use
-#     ` from tensorflow.keras.optimizers.legacy import Adam`
 
 # Author: Romain Tavenard
 # License: BSD 3 clause
 
+import os
+
+# Should be set before importing keras through
+os.environ["KERAS_BACKEND"] = "torch"
+
+from keras.optimizers import Adam
 import numpy
 import matplotlib.pyplot as plt
 
 from tslearn.datasets import CachedDatasets
 from tslearn.preprocessing import TimeSeriesScalerMinMax
 from tslearn.shapelets import LearningShapelets
-
-from tensorflow.keras.optimizers import Adam
-
-# Set a seed to ensure determinism
-numpy.random.seed(42)
 
 # Load the Trace dataset
 X_train, y_train, _, _ = CachedDatasets().load_dataset("Trace")